Patch: remove-slow-messagecache-locking Type: bugfix
Remove locks and hash checking from MessageCache.

Locks were implemented via sleep() (!!!) and led to slow response times when
APC cache was full (it's not expiring anything in that case, just refusing
to store data).

Hash checking is useless because to calculate the right hash for checking
staleness you have to first invalidate the cache :-D. So the hash in fact
was only used as a refresh flag.

Cache update concurrency is still limited to 1 refresh in 60 seconds, but
MessageCache doesn't die when it can't build the cache, it just proceeds
with cache disabled.